Upgrading addon to use both global sandbox setup and the new QUnit hooks API
