+++ addon-test-support/setup-sinon-sandbox.js
@@ -1,0 +1,15 @@
+import { createSandbox, restoreSandbox } from 'ember-sinon-sandbox/test-support/sinon-sandbox';
+
+/**
+ * Allows for creating and restoring a sinon sandbox per test. Uses the *new*
+ * QUnit module API to configure the sandbox on `beforeEach` and `afterEach` of
+ * the test module.
+ *
+ * @export
+ * @param {Object} hooks The QUnit hooks object.
+ * @public
+ */
+export default function setupSinonSandbox(hooks) {
+  hooks.beforeEach(createSandbox);
+  hooks.afterEach(restoreSandbox);
+}
